requirements.txt
pingparsing
jsonpickle
python-crontab